factor out storage providers, update tests
